refactor drafts
