+++ src/app/shared/components/draft-filters/draft-filters.component.html
@@ -1,0 +1,78 @@
+<form class="wrapper-header" [formGroup]="filterGroup">
+  <div class="search-area">
+    <div class="search-bar flex flex-row justify-start items-center">
+      <mat-icon class="search-icon mt-2">search</mat-icon>
+      <input
+        formControlName="searchBarFilter"
+        matInput
+        type="text"
+        class="search"
+        placeholder="{{ 'FORMS.PLACEHOLDERS.SEARCH_BY_TITLE' | translate }}" />
+    </div>
+    <div class="reset" (keydown.enter)="onResetFilters()" (click)="onResetFilters()" *ngIf="!isAreaAdmin">
+      <mat-icon class="reset__icon">refresh</mat-icon>
+      <span class="reset__title">{{ 'HISTORY_LOG.RESET_FILTERS' | translate }}</span>
+    </div>
+  </div>
+  <div class="filters-area">
+    <div class="filter-dropdown">
+      <mat-form-field appearance="fill">
+        <mat-select
+          disableOptionCentering
+          formControlName="workshopDraftStatuses"
+          panelClass="dropdown-panel select-with-form-field"
+          placeholder="{{ 'FORMS.PLACEHOLDERS.SELECT_STATUS' | translate }}"
+          class="select">
+          <mat-option *ngFor="let status of workshopStatusesToFilter" [value]="status" class="dropdown-option">
+            {{ workshopDraftStatusTitles[status] | translate }}
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+    </div>
+    <div class="filter-dropdown">
+      <mat-form-field *ngIf="isTechAdmin || isModerator" appearance="fill">
+        <mat-select
+          disableOptionCentering
+          formControlName="institution"
+          panelClass="dropdown-panel select-with-form-field"
+          placeholder="{{ 'FORMS.PLACEHOLDERS.SELECT_SUBORDINATION' | translate }}"
+          class="select">
+          <mat-option *ngFor="let institution of institutions$ | async" [value]="institution" class="dropdown-option">
+            {{ institution.title }}
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+    </div>
+    <div class="region-filters-area">
+      <div class="filter-dropdown institution-filter">
+        <mat-form-field *ngIf="isTechAdmin || isMinistryAdmin || isModerator" appearance="fill">
+          <mat-select
+            disableOptionCentering
+            formControlName="region"
+            panelClass="dropdown-panel select-with-form-field"
+            placeholder="{{ 'FORMS.PLACEHOLDERS.SELECT_REGION' | translate }}"
+            class="select">
+            <mat-option class="dropdown-option" *ngFor="let region of regions$ | async" [value]="region">
+              {{ region.fullName | translate }}
+            </mat-option>
+          </mat-select>
+        </mat-form-field>
+      </div>
+      <div class="filter-dropdown">
+        <mat-form-field *ngIf="isTechAdmin || isMinistryAdmin || isRegionAdmin || isModerator" appearance="fill">
+          <mat-select
+            [compareWith]="compareCodeficators"
+            disableOptionCentering
+            formControlName="area"
+            panelClass="dropdown-panel select-with-form-field"
+            placeholder="{{ 'FORMS.PLACEHOLDERS.SELECT_TERRITORIAL_COMMUNITY' | translate }}"
+            class="select">
+            <mat-option *ngFor="let codeficator of codeficatorSearch$ | async" [value]="codeficator" class="dropdown-option">
+              {{ codeficator.territorialCommunity }}
+            </mat-option>
+          </mat-select>
+        </mat-form-field>
+      </div>
+    </div>
+  </div>
+</form>
