+++ src/app/shared/components/draft-filters/draft-filters.component.ts
@@ -1,61 +1,42 @@
-import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { AbstractControl, FormBuilder, FormControl, FormGroup } from '@angular/forms';
-import { MatDialog } from '@angular/material/dialog';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
-import { ActivatedRoute } from '@angular/router';
 import { Select, Store } from '@ngxs/store';
-import { debounceTime, distinctUntilChanged, filter, map, Observable, of, skip, startWith, Subject, switchMap, takeUntil } from 'rxjs';
-
-import { Constants, ModeConstants, PaginationConstants } from 'shared/constants/constants';
+import { Observable, takeUntil, switchMap, of, Subject, debounceTime, distinctUntilChanged, filter, map, skip, startWith } from 'rxjs';
+import { PaginationConstants } from 'shared/constants/constants';
 import { CodeficatorCategories } from 'shared/enum/codeficator-categories';
-import { NavBarName } from 'shared/enum/enumUA/navigation-bar';
-import { NoResultsTitle } from 'shared/enum/enumUA/no-results';
-import { OwnershipTypesEnum } from 'shared/enum/enumUA/provider';
-import { DraftStatusEnum, FormOfLearningEnum } from 'shared/enum/enumUA/workshop';
-import { ModalConfirmationType } from 'shared/enum/modal-confirmation';
-import { OwnershipTypes } from 'shared/enum/provider';
+import { DraftStatusEnum } from 'shared/enum/enumUA/workshop';
 import { Role } from 'shared/enum/role';
-import { UserStatusIcons } from 'shared/enum/statuses';
 import { BaseAdmin } from 'shared/models/admin.model';
+import { AreaAdmin } from 'shared/models/area-admin.model';
 import { Codeficator } from 'shared/models/codeficator.model';
+import { CompetitionFilterAdministration } from 'shared/models/competition.model';
 import { Institution } from 'shared/models/institution.model';
 import { PaginationElement } from 'shared/models/pagination-element.model';
 import { RegionAdmin } from 'shared/models/region-admin.model';
-import { SearchResponse } from 'shared/models/search.model';
-import {
-  ApproveCompetitionDraft,
-  GetAreaAdminProfile,
-  GetMinistryAdminProfile,
-  GetRegionAdminProfile,
-  RejectCompetitionDraft
-} from 'shared/store/admin.actions';
+import { WorkshopFilterAdministration } from 'shared/models/workshop.model';
+import { GetMinistryAdminProfile, GetRegionAdminProfile, GetAreaAdminProfile } from 'shared/store/admin.actions';
 import { AdminState } from 'shared/store/admin.state';
 import { FilterState } from 'shared/store/filter.state';
-import { ClearCodeficatorSearch, GetAllInstitutions, GetCodeficatorById, GetCodeficatorSearch } from 'shared/store/meta-data.actions';
+import { GetAllInstitutions, GetCodeficatorSearch, GetCodeficatorById, ClearCodeficatorSearch } from 'shared/store/meta-data.actions';
 import { MetaDataState } from 'shared/store/meta-data.state';
-import { PopNavPath, PushNavPath } from 'shared/store/navigation.actions';
 import { GetProfile } from 'shared/store/registration.actions';
 import { RegistrationState } from 'shared/store/registration.state';
-import { Util } from 'shared/utils/utils';
-import { WorkshopDraftStatus } from 'shared/enum/workshop';
-import { Competition, CompetitionDraft, CompetitionFilterAdministration } from 'shared/models/competition.model';
-import { ReasonModalWindowComponent } from '../confirmation-modal-window/reason-modal-window/reason-modal-window.component';
 
 @Component({
-  selector: 'app-competition-list',
-  templateUrl: './competition-list.component.html',
-  styleUrls: ['./competition-list.component.scss']
+  selector: 'app-draft-filters',
+  templateUrl: './draft-filters.component.html',
+  styleUrl: './draft-filters.component.scss'
 })
-export class CompetitionListComponent implements OnInit, OnDestroy {
-  @Input() public setCompetitionFiltersByDefault: (
-    competitionParameters: CompetitionFilterAdministration,
-    role: Role,
-    selectedAdmin: BaseAdmin
-  ) => void;
-
-  @Output() public getWorkshopsByFilter: EventEmitter<CompetitionFilterAdministration> = new EventEmitter();
-
+export class DraftFiltersComponent implements OnInit, OnDestroy {
+  @Input() public currentPage: PaginationElement = PaginationConstants.firstPage;
+  @Output() public getDraftsByFilter: EventEmitter<{
+    parameters: CompetitionFilterAdministration | WorkshopFilterAdministration;
+    currentPage: PaginationElement;
+  }> = new EventEmitter();
+  @Select(RegistrationState.role)
+  public role$: Observable<Role>;
+  @Select(AdminState.selectedAdmin)
+  public selectedAdmin$: Observable<BaseAdmin>;
   @Select(AdminState.isLoading)
   public isLoadingCabinet$: Observable<boolean>;
   @Select(MetaDataState.institutions)
@@ -64,52 +45,21 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
   public settlement$: Observable<Codeficator>;
   @Select(MetaDataState.codeficatorSearch)
   public codeficatorSearch$: Observable<Codeficator[]>;
-  @Select(RegistrationState.role)
-  public role$: Observable<Role>;
-  @Select(AdminState.selectedAdmin)
-  public selectedAdmin$: Observable<BaseAdmin>;
 
-  public readonly noWorkshops = NoResultsTitle.noResult;
-  public readonly modeConstants = ModeConstants;
-  public readonly tooltipPosition = Constants.MAT_TOOL_TIP_POSITION_BELOW;
-  public readonly ownershipTypeEnum = OwnershipTypesEnum;
-  public readonly formOfLearningEnum = FormOfLearningEnum;
-  public readonly ownershipTypes = OwnershipTypes;
-  public readonly statusIcons = UserStatusIcons;
-  public readonly UNLIMITED_SEATS = Constants.UNLIMITED_SEATS;
-  public readonly workshopDraftStatus = WorkshopDraftStatus;
   public readonly workshopDraftStatusTitles = DraftStatusEnum;
-  public readonly workshopStatusesToFilter = ['PendingModeration', 'EditedByModerator'];
-  public readonly displayedColumns: string[] = [
-    'title',
-    // 'formOfLearning',
-    // 'seats',
-    'providerEdrpou',
-    'directorPosition',
-    'directorFullName',
-    'isPaid',
-    'status',
-    'actions'
-  ];
-  public competitionParameters: CompetitionFilterAdministration = {};
-  public dataSource = new MatTableDataSource<CompetitionDraft>();
-  public currentPage: PaginationElement = PaginationConstants.firstPage;
-
   public selectedAdmin: BaseAdmin;
-  public role: Role;
-  public competition: Competition;
-  public selectedCompetitionDraftId: string;
-  public isInfoDisplayed: boolean;
   public filterGroup: FormGroup;
-  public totalEntities: number;
+  public role: Role;
+
+  // TO DO Make one parameters object for both competitions and workshops
+  public parameters: CompetitionFilterAdministration & WorkshopFilterAdministration = {};
   public regions$: Observable<Codeficator[]>;
 
+  public readonly workshopStatusesToFilter = ['PendingModeration', 'EditedByModerator'];
   private readonly destroy$: Subject<void> = new Subject<void>();
 
   constructor(
-    protected readonly route: ActivatedRoute,
     private readonly store: Store,
-    private readonly matDialog: MatDialog,
     private readonly formBuilder: FormBuilder
   ) {}
 
@@ -153,16 +103,6 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
     return this.filterGroup.get('workshopDraftStatuses') as FormControl;
   }
 
-  @Input()
-  public set competitions(value: SearchResponse<CompetitionDraft[]>) {
-    this.dataSource.data = value?.entities;
-    this.totalEntities = value?.totalAmount;
-  }
-
-  public compareCodeficators(codeficator1: Codeficator, codeficator2: Codeficator): boolean {
-    return codeficator1.id === codeficator2.id;
-  }
-
   public ngOnInit(): void {
     this.selectedAdmin$.pipe(takeUntil(this.destroy$)).subscribe((admin: BaseAdmin) => (this.selectedAdmin = admin));
 
@@ -189,16 +129,8 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
       )
       .subscribe(() => {
         this.setInitialWorkshopFilterByDefault();
-        this.getWorkshops();
+        this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
       });
-
-    this.store.dispatch(
-      new PushNavPath({
-        name: NavBarName.CompetitionDrafts,
-        isActive: false,
-        disable: true
-      })
-    );
 
     this.filterGroup = this.formBuilder.group({
       searchBarFilter: new FormControl(''),
@@ -212,36 +144,35 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
     this.subscribeFormControls();
   }
 
-  public onViewCompetitionInfo(competition: CompetitionDraft): void {
-    this.selectedCompetitionDraftId = competition.competitiveEventDraftId;
-    this.competition = competition.competitiveEventDetails;
-    this.isInfoDisplayed = true;
+  public compareCodeficators(codeficator1: Codeficator, codeficator2: Codeficator): boolean {
+    return codeficator1.id === codeficator2.id;
   }
 
-  public onRejectDraft(competition: CompetitionDraft): void {
-    const dialogRef = this.matDialog.open(ReasonModalWindowComponent, {
-      data: { type: ModalConfirmationType.editingCompetition }
-    });
-    dialogRef
-      .afterClosed()
-      .pipe(filter(Boolean))
-      .subscribe((statusReason: string) =>
-        this.store.dispatch(new RejectCompetitionDraft(competition.competitiveEventDraftId, statusReason))
-      );
+  public ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.unsubscribe();
   }
 
-  public onApproveDraft(competition: CompetitionDraft): void {
-    this.store.dispatch(new ApproveCompetitionDraft(competition.competitiveEventDraftId));
-  }
+  public setFiltersByDefault(parameters: CompetitionFilterAdministration, role: Role, selectedAdmin?: BaseAdmin): void {
+    parameters.searchString = '';
+    parameters.size = PaginationConstants.TABLE_ITEMS_PER_PAGE;
 
-  public onPageChange(page: PaginationElement): void {
-    this.currentPage = page;
-    this.getWorkshops();
-  }
-
-  public onItemsPerPageChange(itemsPerPage: number): void {
-    this.competitionParameters.size = itemsPerPage;
-    this.onPageChange(PaginationConstants.firstPage);
+    switch (role) {
+      case Role.techAdmin:
+      case Role.moderator:
+        parameters.institutionId = '';
+        parameters.catottgId = 0;
+        break;
+      case Role.ministryAdmin:
+        parameters.institutionId = selectedAdmin.institutionId;
+        parameters.catottgId = 0;
+        break;
+      case Role.regionAdmin:
+      case Role.areaAdmin:
+        parameters.institutionId = selectedAdmin.institutionId;
+        parameters.catottgId = (selectedAdmin as RegionAdmin | AreaAdmin).catottgId;
+        break;
+    }
   }
 
   public onResetFilters(): void {
@@ -259,19 +190,8 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
       }
 
       this.setInitialWorkshopFilterByDefault();
-      this.getWorkshops();
+      this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
     }
-  }
-
-  public closeInfo(): void {
-    this.isInfoDisplayed = false;
-    this.selectedCompetitionDraftId = null;
-  }
-
-  public ngOnDestroy(): void {
-    this.destroy$.next();
-    this.destroy$.unsubscribe();
-    this.store.dispatch(new PopNavPath());
   }
 
   private setInformationDependingOnRole(): void {
@@ -325,25 +245,25 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
         takeUntil(this.destroy$)
       )
       .subscribe((searchValue: string) => {
-        this.competitionParameters.searchString = searchValue;
+        this.parameters.searchString = searchValue;
         this.currentPage = PaginationConstants.firstPage;
-        this.getWorkshops();
+        this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
       });
 
     this.institutionFormControl.valueChanges
       .pipe(distinctUntilChanged(), startWith(''), skip(1), debounceTime(1000), filter(Boolean), takeUntil(this.destroy$))
       .subscribe(() => {
-        this.competitionParameters.institutionId = this.institutionFormControl.value.id;
+        this.parameters.institutionId = this.institutionFormControl.value.id;
         this.currentPage = PaginationConstants.firstPage;
-        this.getWorkshops();
+        this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
       });
 
     this.regionFormControl.valueChanges
       .pipe(distinctUntilChanged(), startWith(''), skip(1), debounceTime(1000), filter(Boolean), takeUntil(this.destroy$))
       .subscribe((value: Codeficator) => {
-        this.competitionParameters.catottgId = this.regionFormControl.value.id;
+        this.parameters.catottgId = this.regionFormControl.value.id;
         this.currentPage = PaginationConstants.firstPage;
-        this.getWorkshops();
+        this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
         if (value.category === CodeficatorCategories.Region) {
           this.store.dispatch(new GetCodeficatorSearch('', [CodeficatorCategories.TerritorialCommunity], this.regionFormControl.value.id));
           this.areaFormControl.enable();
@@ -356,26 +276,23 @@ export class CompetitionListComponent implements OnInit, OnDestroy {
     this.areaFormControl.valueChanges
       .pipe(distinctUntilChanged(), startWith(''), skip(1), debounceTime(1000), filter(Boolean), takeUntil(this.destroy$))
       .subscribe(() => {
-        this.competitionParameters.catottgId = this.areaFormControl.value.id;
+        this.parameters.catottgId = this.areaFormControl.value.id;
         this.currentPage = PaginationConstants.firstPage;
-        this.getWorkshops();
+        this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
       });
 
     this.statusFormControl.valueChanges.pipe(distinctUntilChanged(), debounceTime(1000), takeUntil(this.destroy$)).subscribe(() => {
-      this.competitionParameters.competitiveEventDraftStatuses = this.statusFormControl.value;
+      // TO DO Change statuses when it will be unified
+      this.parameters.competitiveEventDraftStatuses = this.statusFormControl.value;
+      this.parameters.workshopDraftStatuses = this.statusFormControl.value;
       this.currentPage = PaginationConstants.firstPage;
-      this.getWorkshops();
+      this.getDraftsByFilter.emit({ parameters: this.parameters, currentPage: this.currentPage });
     });
   }
 
   private setInitialWorkshopFilterByDefault(): void {
-    this.competitionParameters.searchString = '';
-    this.competitionParameters.size = PaginationConstants.TABLE_ITEMS_PER_PAGE;
-    this.setCompetitionFiltersByDefault(this.competitionParameters, this.role, this.selectedAdmin);
-  }
-
-  private getWorkshops(): void {
-    Util.setFromPaginationParam(this.competitionParameters, this.currentPage, this.totalEntities);
-    this.getWorkshopsByFilter.emit(this.competitionParameters);
+    this.parameters.searchString = '';
+    this.parameters.size = PaginationConstants.TABLE_ITEMS_PER_PAGE;
+    this.setFiltersByDefault(this.parameters, this.role, this.selectedAdmin);
   }
 }
