--- src/app/shared/components/workshop-list/workshop-list.component.scss
@@ -1,213 +1,0 @@
-@use 'tables';
-@use 'list-wrappers';
-@use 'search';
-@use 'dropdown';
-
-.total-count {
-  font-family: Innerspace, sans-serif;
-  font-size: 14px;
-  font-weight: bold;
-  color: white;
-}
-
-.workshop-list-status {
-  font-family: 'Open Sans', sans-serif;
-  font-weight: 600;
-  font-size: 13px;
-  margin: 5px;
-}
-
-.Approved {
-  color: var(--status-approved);
-}
-
-.PendingModeration {
-  color: var(--status-pending);
-}
-
-.EditedByModerator,
-.Recheck {
-  color: var(--status-recheck);
-}
-
-.Blocked {
-  color: var(--status-blocked);
-}
-
-.NotProvided {
-  color: var(--status-not-provided);
-}
-
-.wrapper-workshop-list {
-  position: relative;
-  margin-top: 2rem;
-}
-
-.container {
-  margin-left: 1rem;
-  @media (max-width: 676px) {
-    margin-left: 0;
-  }
-}
-
-.search {
-  padding-left: 10px;
-  background-color: transparent;
-  color: var(--primary-dark-text-color);
-}
-.search-area,
-.filters-area,
-.region-filters-area {
-  display: flex;
-  justify-content: flex-end;
-  z-index: 10000;
-  position: static;
-  right: 12px;
-  align-items: center;
-}
-
-.search-icon {
-  transform: rotate(90deg);
-}
-
-.status-title {
-  padding-left: 7px;
-}
-
-.true {
-  color: var(--status-approved);
-}
-
-.false {
-  color: var(--status-pending);
-}
-
-.wrapper-workshop-list-container {
-  display: flex;
-  flex-direction: row;
-}
-
-.table-container {
-  height: fit-content;
-}
-
-.select {
-  min-width: 100px;
-  background: var(--bg-specific-light);
-  border-radius: 20px;
-  box-sizing: border-box;
-  box-shadow: 0 6px 16px rgba(0, 0, 0, 0.08);
-  overflow: visible;
-  padding: 0.5rem 1rem;
-}
-
-.select,
-.options {
-  font-family: 'Open Sans', sans-serif;
-  font-style: normal;
-  font-size: 13px;
-  line-height: 22px;
-  height: 40px;
-  color: var(--primary-dark-text-color) !important;
-}
-
-.filters-area,
-.region-filters-area {
-  column-gap: 5px;
-  margin-right: 12px;
-}
-
-.reset {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  cursor: pointer;
-  margin-left: 15px;
-
-  &:hover {
-    opacity: 0.5;
-  }
-
-  &__icon {
-    color: var(--primary-light-text-color) !important;
-    font-size: 24px;
-    transform: rotateY(180deg) rotate(45deg);
-    overflow: visible;
-  }
-
-  &__title {
-    font-size: 0.8125rem;
-    line-height: 19px;
-    color: var(--primary-light-text-color);
-  }
-}
-
-@media (max-width: 1245px) {
-  .reset {
-    &__title {
-      display: none;
-    }
-  }
-}
-
-@media (max-width: 1415px) {
-  .wrapper-header {
-    flex-direction: column;
-  }
-
-  .search-area {
-    width: 100%;
-    position: relative;
-    right: 0;
-  }
-
-  .search-bar {
-    width: 100%;
-  }
-}
-
-@media (max-width: 925px) {
-  .filters-area {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-
-  .region-filters-area {
-    margin-right: 0;
-    margin-left: 10px;
-    margin-top: 0 !important;
-  }
-
-  .filter-dropdown {
-    margin-top: 10px;
-  }
-
-  .select {
-    margin-left: 0;
-  }
-}
-
-td {
-  word-break: break-word;
-}
-
-.tab-text {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-::ng-deep .filters-area .mdc-text-field {
-  background-color: transparent;
-  padding: 0;
-}
-
-.reset-title {
-  color: var(--primary-light-text-color);
-}
-
-::ng-deep .mat-mdc-select-placeholder {
-  color: var(--primary-dark-text-color) !important;
-}
