+++ src/app/shell/admin-tools/data/admin-competition-list/competition-info/competition-info.component.ts
