+++ src/app/shell/admin-tools/data/admin-workshop-list/workshop-info/workshop-info.component.spec.ts
