+++ src/app/shell/personal-cabinet/provider/create-workshop/create-description-form/create-description-form.component.html
@@ -3,7 +3,7 @@
     *ngIf="isImagesFeature"
     [imgMaxAmount]="10"
     [cropperConfig]="cropperConfig"
-    [label]="'Фотографії'"
+    [label]="'FORMS.LABELS.PHOTOS' | translate"
     formControlName="imageFiles"
     [imageIdsFormControl]="DescriptionFormGroup.get('imageIds')"
     (deleteImage)="onImageDelete()">
